Update test_pow for shorter block times
